fix: getFileType to utils
